Seperate classes form from main.rb
