+++ student.rb
@@ -1,0 +1,11 @@
+require_relative 'person'
+class Student < Person
+  def initialize(age, classroom, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission)
+    @classroom = classroom
+  end
+
+  def play_hook
+    "¯\(ツ)/¯"
+  end
+end
